+++ Cargo.toml
@@ -10,12 +10,11 @@ documentation = "http://mrhooray.github.io/crc-rs/crc/index.html"
 license = "MIT"
 
 [dependencies]
-
-[dependencies.build_const]
-version = "0.2"
-default-features = false
+util = { path = "./src/util" }
+build_const = { version = "0.2", default-features = false }
 
 [build-dependencies]
+util = { path = "./src/util" }
 build_const = "0.2"
 
 [features]
