+++ src/crc64.rs
@@ -3,6 +3,8 @@ use std::hash::Hasher;
 #[cfg(not(feature = "std"))]
 use core::hash::Hasher;
 
+pub use util::make_table_crc64 as make_table;
+
 build_const!("crc64_constants");
 
 pub struct Digest {
@@ -15,22 +17,6 @@ pub trait Hasher64 {
     fn reset(&mut self);
     fn write(&mut self, bytes: &[u8]);
     fn sum64(&self) -> u64;
-}
-
-pub fn make_table(poly: u64) -> [u64; 256] {
-    let mut table = [0u64; 256];
-    for i in 0..256 {
-        let mut value = i as u64;
-        for _ in 0..8 {
-            value = if (value & 1) == 1 {
-                (value >> 1) ^ poly
-            } else {
-                value >> 1
-            }
-        }
-        table[i] = value;
-    }
-    table
 }
 
 pub fn update(mut value: u64, table: &[u64; 256], bytes: &[u8]) -> u64 {
