+++ src/util/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "util"
+version = "0.1.0"
+
+[lib]
+name = "util"
+path = "lib.rs"
